+++ nav2_route/src/goal_intent_extractor.cpp
@@ -21,7 +21,6 @@
 #include "nav2_costmap_2d/cost_values.hpp"
 #include "nav2_costmap_2d/costmap_2d.hpp"
 #include "nav2_costmap_2d/costmap_subscriber.hpp"
-#include "nav2_route/breadth_first_search.hpp"
 #include "nav2_route/goal_intent_extractor.hpp"
 #include "nav2_util/node_utils.hpp"
 #include "nav_msgs/msg/occupancy_grid.hpp"
